+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Devel
 
 - Bugfix: require records with `access=OWNER` to have `can_compute=True` for `WorkspaceGroupSharing` objects.
-- Add information about deleting GroupAccountMembership, GrooupGroupMembership, and WorkspaceGroupSharing records to the documtation User guide.
+- Add information about deleting GroupAccountMembership, GrooupGroupMembership, and WorkspaceGroupSharing records to the documentation User guide.
+- Add the `run_anvil_audit` management command to run AnVIL audits.
 
 ## 0.12.1 (2023-02-16)